feat: load config from a file; refactor CotProject (#122)
This makes a lot of changes to how the `CotProject` (now just `Project`)
is bootstrapped. This transfers much of the control to CliTasks, and
hopefully will be more scalable and clean.

This also removes the `Send` bound from CliTasks, since it doesn't look
like it's necessary.

Making the config structure serializable again makes it easy to display
it on the error pages (which this commit implements, too).