fix: add Cargo.lock to project template to avoid broken dependencies (#191)
By pinning new project's dependencies to specific versions, we avoid
problems like #187 which are caused by changes in the dependencies
(in this case we were using a prerelease version of a dependency which
updated one of its dependencies to edition=2024, but this will also
prevent accidental semver-breaking changes from breaking our builds,
too).

This also adds a justfile that makes it easy to update the lockfile.