feat!: use extractor pattern for request handlers (#253)
This is a massive change and the transition probably isn't even fully
done yet. This should be a good starting point, though.

The remaining work includes:

* Removing more of the `RequestExt` trait
* Creating `IntoResponse` trait so that the consumers can return other
  types than `cot::Result<Response>`
* More work on maintaining consistency in the API

Note that this is a (quite big) breaking change, mostly because it
removes a lot of methods from the `RequestExt` trait.

This change is needed to implement automatic OpenAPI spec generation
in #159.

This also includes some other changes:

feat: add `RegisterAppsContext`, `AuthBackendContext`, `MiddlewareContext`
as type aliases for `ProjectContext` in specific bootstrapping phases
that are more semantic and whose names won't change when changing
the phases

chore!: Remove methods from the `RequestExt` that duplicate extractors'
functionalities

feat: display object paths when (de)serialization error happened with serde

chore!: `AuthRequestExt` trait is now replaced by `Auth` struct and
`AuthMiddleware` is now required for authentication to work

chore!: add `WithDatabase` bootstrapping phase

chore: `Urls` object can now be used with the `reverse!` macro and
not only `Request`