fix(cli): migration generator not working in inner project directories (#267)
This also fixes `--app-name` param trying to find existing packages (its
original use is to override the app name even when it does not
correspond to any existing crate).

The change also makes `migration make` not work when ran in a workspace,
instead forcing the user to choose a specific crate inside. In the
future, a proper support should be added instead (possibly taking
dependencies between crates into account) and an option to specify a
crate name without having to specify its path.